:sparkles: add version in CLI
