+++ setup.cfg
@@ -34,6 +34,10 @@ install_requires =
     pytz>=2022.1
     requests~=2.28
 
+[options.entry_points]
+console_scripts =
+    mindee-cli = mindee.cli:main
+
 [options.package_data]
 mindee =
     version
